refactor: standardize admin api into op-service
