+++ op-node/node/api.go
@@ -11,7 +11,7 @@ import (
 	"github.com/ethereum-optimism/optimism/op-node/rollup"
 	"github.com/ethereum-optimism/optimism/op-node/version"
 	"github.com/ethereum-optimism/optimism/op-service/eth"
-	oplog "github.com/ethereum-optimism/optimism/op-service/log"
+	"github.com/ethereum-optimism/optimism/op-service/rpc"
 )
 
 type l2EthClient interface {
@@ -31,68 +31,40 @@ type driverClient interface {
 	SequencerActive(context.Context) (bool, error)
 }
 
-type rpcMetrics interface {
-	// RecordRPCServerRequest returns a function that records the duration of serving the given RPC method
-	RecordRPCServerRequest(method string) func()
-}
-
 type adminAPI struct {
-	dr  driverClient
-	m   rpcMetrics
-	log log.Logger
+	*rpc.CommonAdminAPI
+	dr driverClient
 }
 
-func NewAdminAPI(dr driverClient, m rpcMetrics, log log.Logger) *adminAPI {
+func NewAdminAPI(dr driverClient, m rpc.RpcMetrics, log log.Logger) *adminAPI {
 	return &adminAPI{
-		dr:  dr,
-		m:   m,
-		log: log,
+		dr:             dr,
+		CommonAdminAPI: rpc.NewCommonAdminAPI(m, log),
 	}
 }
 
 func (n *adminAPI) ResetDerivationPipeline(ctx context.Context) error {
-	recordDur := n.m.RecordRPCServerRequest("admin_resetDerivationPipeline")
+	recordDur := n.M.RecordRPCServerRequest("admin_resetDerivationPipeline")
 	defer recordDur()
 	return n.dr.ResetDerivationPipeline(ctx)
 }
 
 func (n *adminAPI) StartSequencer(ctx context.Context, blockHash common.Hash) error {
-	recordDur := n.m.RecordRPCServerRequest("admin_startSequencer")
+	recordDur := n.M.RecordRPCServerRequest("admin_startSequencer")
 	defer recordDur()
 	return n.dr.StartSequencer(ctx, blockHash)
 }
 
 func (n *adminAPI) StopSequencer(ctx context.Context) (common.Hash, error) {
-	recordDur := n.m.RecordRPCServerRequest("admin_stopSequencer")
+	recordDur := n.M.RecordRPCServerRequest("admin_stopSequencer")
 	defer recordDur()
 	return n.dr.StopSequencer(ctx)
 }
 
 func (n *adminAPI) SequencerActive(ctx context.Context) (bool, error) {
-	recordDur := n.m.RecordRPCServerRequest("admin_sequencerActive")
+	recordDur := n.M.RecordRPCServerRequest("admin_sequencerActive")
 	defer recordDur()
 	return n.dr.SequencerActive(ctx)
-}
-
-func (n *adminAPI) SetLogLevel(ctx context.Context, lvlStr string) error {
-	recordDur := n.m.RecordRPCServerRequest("admin_setLogLevel")
-	defer recordDur()
-
-	h := n.log.GetHandler()
-
-	lvl, err := log.LvlFromString(lvlStr)
-	if err != nil {
-		return err
-	}
-
-	// We set the log level, and do not wrap the handler with an additional filter handler,
-	// as the underlying handler would otherwise also still filter with the previous log level.
-	lvlSetter, ok := h.(oplog.LvlSetter)
-	if !ok {
-		return fmt.Errorf("log handler type %T cannot change log level", h)
-	}
-	lvlSetter.SetLogLevel(lvl)
-	return nil
 }
 
 type nodeAPI struct {
@@ -100,10 +72,10 @@ type nodeAPI struct {
 	client l2EthClient
 	dr     driverClient
 	log    log.Logger
-	m      rpcMetrics
+	m      rpc.RpcMetrics
 }
 
-func NewNodeAPI(config *rollup.Config, l2Client l2EthClient, dr driverClient, log log.Logger, m rpcMetrics) *nodeAPI {
+func NewNodeAPI(config *rollup.Config, l2Client l2EthClient, dr driverClient, log log.Logger, m rpc.RpcMetrics) *nodeAPI {
 	return &nodeAPI{
 		config: config,
 		client: l2Client,
